services/xml-service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pandas
lxml
aiohttp
python-dotenv
python-multipart